0.8.4/requirements.txt
python-dotenv==1.1.1
web3==7.12.1

